split files and include test for signature
