+++ .gitignore
@@ -9,6 +9,7 @@
 *.userosscache
 *.sln.docstates
 
+/vendor/
 # User-specific files (MonoDevelop/Xamarin Studio)
 *.userprefs
