+++ composer.json
@@ -3,6 +3,10 @@
   "type": "library",
   "description": "The Official Smile Identity PHP library",
   "require": {
+    "php": ">=7.4 <8.0",
     "maennchen/zipstream-php": "^1.2"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^9.5"
   }
 }
