+++ lib/SamplePhpOop.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace model;
+require 'config.php';
+
+class SampleClass
+{
+    public function getName()
+    {
+        echo DEFAULT_JOB_STATUS_TIMEOUT;
+    }
+}
+
+$sampleClass = new SampleClass;
+$sampleClass->getName();
