+++ lib/config.php
@@ -1,0 +1,9 @@
+<?php
+
+define('VERSION', '1.1.0');
+define('DEFAULT_JOB_STATUS_TIMEOUT', 20);
+define('DEFAULT_JOB_STATUS_SLEEP', 2);
+define('SID_SERVERS', array(
+    'https://3eydmgh10d.execute-api.us-west-2.amazonaws.com/test',
+    'https://la7am6gdm8.execute-api.us-west-2.amazonaws.com/prod'
+));
